Moved Kind to base package 'javaslang' + javadoc
