+++ src-gen/main/java/javaslang/algebra/CheckedMonad.java
@@ -10,6 +10,7 @@ package javaslang.algebra;
 \*-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-*/
 
 import java.util.function.Consumer;
+import javaslang.Kind;
 import javaslang.control.Try.CheckedConsumer;
 import javaslang.control.Try.CheckedFunction;
 import javaslang.control.Try.CheckedPredicate;
@@ -25,7 +26,7 @@ import javaslang.control.Try.CheckedPredicate;
  * </ul>
  * given
  * <ul>
- * <li>an object {@code m} of type {@code HigherKinded<T, M>}</li>
+ * <li>an object {@code m} of type {@code Kind<M, T>}</li>
  * <li>an object {@code a} of type T</li>
  * <li>a constructor {@code unit} taking an {@code a} and producing an object of type {@code M}</li>
  * <li>a function {@code f: T → M}
