+++ src/main/java/javaslang/algebra/package-info.java
@@ -1,6 +1,6 @@
 /**
  * <p>Algebraic interfaces used to describe inherent class properties like {@linkplain javaslang.algebra.Monad}.</p>
- * <p>Interfaces like {@linkplain javaslang.algebra.Kind} are needed to enhance API design and are mainly used
+ * <p>Interfaces like {@linkplain javaslang.Kind} are needed to enhance API design and are mainly used
  * internally by Javaslang. An example for {@linkplain javaslang.algebra.Monoid} usage is
  * {@link javaslang.collection.Traversable#foldMap(javaslang.algebra.Monoid, java.util.function.Function)}.</p>
  *
